requirements.txt
flask==3.1.1
openai>=2.0.0
gunicorn==23.0.0
lark-oapi>=1.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
